requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pyyaml==6.0.1
httpx==0.25.2
pdfplumber==0.10.3
pymupdf==1.23.8
nomic==1.1.9
chromadb==0.4.18
faiss-cpu==1.7.4
numpy==1.24.3
python-multipart==0.0.6
